+++ CHANGELOG.md
@@ -6,10 +6,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Support renaming profiles
+
 ### Changed
 - Profile name is no longer displayed if default
 - Trying to create a profile that already exists throws an error
 - Trying to delete a profile that is currently in use throws an error
+- Use argparse for argument parsing instead of getopt
+
+### Removed
+- Removed manpage
 
 ## [0.11.1] - 2017-12-19
 ### Packaging